Migrate to a more modern bundling pipeline
- babel@6
- babel-preset-env
- webtask@2
- Use the 'all npm' feature for external modules
