+++ jsconfig.json
@@ -3,6 +3,6 @@
         "target": "ES6"
     },
     "exclude": [
-        "/node_modules/"
+        "node_modules"
     ]
 }
