+++ lib/compiler.js
@@ -5,20 +5,22 @@ const Rx = require('rxjs');
 const Webpack = require('webpack');
 
 
-module.exports = compileBundle;
+module.exports = {
+    compile,
+};
 
 
-function compileBundle(options) {
+function compile(options) {
     const memFs = new MemoryFs();
     const compiler = Webpack(options.webpackConfig);
     let generation = 0;
-    
+
     compiler.outputFileSystem = memFs;
-    
+
     return Rx.Observable.create(function (subscriber) {
         if (options.watch) {
             const watcher = compiler.watch({}, onWatch);
-            
+
             // Return an async disposer function
             return Rx.Observable.defer(function () {
                 return Rx.Observable.bindNodeCallback(watcher.close.bind(watcher));
@@ -26,14 +28,14 @@ function compileBundle(options) {
         } else {
             compiler.run(onRun);
         }
-        
+
         function onGeneration(stats) {
             try {
                 const info = stats.toJson();
                 const code = stats.hasErrors()
                     ?   undefined
                     :   memFs.readFileSync('/bundle.js', 'utf8');
-                
+
                 subscriber.next({
                     code: code,
                     generation: ++generation,
@@ -43,28 +45,27 @@ function compileBundle(options) {
             } catch (e) {
                 subscriber.error(e);
             }
-            
         }
-        
+
         function onRun(err, stats) {
             if (err) {
                 console.warn(err.message);
-                
+
                 return subscriber.error(err);
             }
-            
+
             onGeneration(stats);
-            
+
             subscriber.complete();
         }
-        
+
         function onWatch(err, stats) {
             if (err) {
                 console.warn(err.message);
-                
+
                 return subscriber.error(err);
             }
-            
+
             onGeneration(stats);
         }
     });
