+++ lib/helpers.js
@@ -32,20 +32,20 @@ function getDirname (entry) {
     const segments = Path.dirname(Path.normalize(entry))
         .split(Path.sep)
         .filter(Boolean);
-    
+
     while (segments.length) {
         try {
             const pathname = Path.sep + Path.join.apply(null, segments.concat(['package.json']));
-            
+
             // When require doesn't throw, we break out of the loop
             require(pathname);
 
             break;
         } catch (__) { }
-        
+
         segments.pop();
     }
-    
+
     return segments.length
         ?   Path.resolve(Path.sep + Path.join.apply(null, segments))
         :   process.cwd();
@@ -56,15 +56,18 @@ function getPackageJsonPath(entry) {
 }
 
 function loadBuiltins(cb) {
+    // return Superagent.post(
+
+
     Superagent.get(LIST_MODULES_URL)
         .accept('json')
         .end(function (err, res) {
             if (err) return cb(err);
-            
+
             if (!res.body || !Array.isArray(res.body.modules)) {
                 return cb(new Error('Unexpected response format when requesting webtask modules'));
             }
-            
+
             var modules = res.body.modules.reduce(function (acc, module) {
                 if (module.version === 'native') {
                     acc.native[module.name] = true;
@@ -72,16 +75,16 @@ function loadBuiltins(cb) {
                     if (!acc.installed[module.name]) {
                         acc.installed[module.name] = [];
                     }
-                    
+
                     acc.installed[module.name].push(module.version);
                 }
-                
+
                 return acc;
             }, {
                 native: {},
                 installed: BUILTIN,
             });
-            
+
             return cb(null, modules);
         });
 }
