+++ package.json
@@ -20,24 +20,21 @@
   "dependencies": {
     "argparse": "^1.0.7",
     "async": "^2.1.4",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.8",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-runtime": "^6.18.0",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-preset-env": "^1.3.3",
     "chalk": "^1.1.3",
     "cli-table": "^0.3.1",
     "glob": "^6.0.4",
-    "json-loader": "^0.5.4",
     "lodash": "^4.9.0",
     "memory-fs": "^0.3.0",
     "mkdirp": "^0.5.1",
     "raw-loader": "^0.5.1",
     "readline-sync": "^1.3.1",
-    "rxjs": "^5.0.0-rc.4",
+    "rxjs": "^5.3.0",
     "semver": "^5.1.0",
     "superagent": "^1.7.2",
-    "webpack": "^1.13.1"
+    "webpack": "^2.3.3"
   },
   "devDependencies": {
     "code": "^4.0.0",
